Update victorian.sql
The USE srve222 command at the top of the file tells us what database to put the tables into. The commands create two tables, book and comments. These tables are used to store the information about the book as well as the comments that people give about the book.When the user searches for something, the database is accessed and the book table is searched through with the search query being changed depending on what is being search for. The comments table is added to when somebody adds a new comment about a certain book. 

The comments table includes a foreign key into the book table accessing the bid that way the two tables can easily be connected. in order to show the correct comments on a certain books webpage.